modules/private-endpoint/scripts/requirements.txt
dnspython
wrapt-timeout-decorator